requirements.txt
numpy
pandas
ipykernel
pydantic
python-dotenv
langchain
langchain-openai
langchain-groq
langchain-core
langchain-community
defusedxml